:construction: 结构重构

+++ galaxy-web-app/galaxy-oj/src/main/java/io/charlie/web/oj/modular/sys/role/entity/SysRole.java
@@ -40,5 +40,5 @@ public class SysRole extends CommonEntity {
     private String description;
 
     @Schema(description = "角色层级")
-    private Byte level;
+    private Integer level;
 }
